refactor: new method for local testing
